Block topkickstarterprojects.com spam
Got unsolicited email from this domain. All of these domains were linked
in the email, with the exception of trendsender.com. When I searched for
the from email, one of the trendsender.com pages showed an email
to a topkickstarterprojects.com address as the POC (for reporting spam,
no less!) Assuming it's affiliated with the TKP spam